+++ project_backend/src/main/java/ru/stray27/project_backend/dto/ResourceDto.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ResourceDto {
-    private Long id;
+    private Integer id;
     private Integer amount;
-    private Long settlementId;
+    private Integer settlementId;
 }
