+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/AnimalRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Animal;
 
-public interface AnimalRepository extends CrudRepository<Animal, Long> {
+public interface AnimalRepository extends CrudRepository<Animal, Integer> {
 }
