+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/BattleRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Battle;
 
-public interface BattleRepository extends CrudRepository<Battle, Long> {
+public interface BattleRepository extends CrudRepository<Battle, Integer> {
 }
