+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/CasteRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Caste;
 
-public interface CasteRepository extends CrudRepository<Caste, Long> {
+public interface CasteRepository extends CrudRepository<Caste, Integer> {
 }
