+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/EnchantmentRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Enchantment;
 
-public interface EnchantmentRepository extends CrudRepository<Enchantment, Long> {
+public interface EnchantmentRepository extends CrudRepository<Enchantment, Integer> {
 }
