+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/MonsterRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Monster;
 
-public interface MonsterRepository extends CrudRepository<Monster, Long> {
+public interface MonsterRepository extends CrudRepository<Monster, Integer> {
 }
