+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/PeopleRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.People;
 
-public interface PeopleRepository extends CrudRepository<People, Long> {
+public interface PeopleRepository extends CrudRepository<People, Integer> {
 }
