+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/ResourceRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Resource;
 
-public interface ResourceRepository extends CrudRepository<Resource, Long> {
+public interface ResourceRepository extends CrudRepository<Resource, Integer> {
 }
