+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/WeaponRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.Weapon;
 
-public interface WeaponRepository extends CrudRepository<Weapon, Long> {
+public interface WeaponRepository extends CrudRepository<Weapon, Integer> {
 }
