+++ project_backend/src/main/java/ru/stray27/project_backend/repositories/WorldRepository.java
@@ -3,5 +3,5 @@ package ru.stray27.project_backend.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.stray27.project_backend.entities.World;
 
-public interface WorldRepository extends CrudRepository<World, Long> {
+public interface WorldRepository extends CrudRepository<World, Integer> {
 }
